% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c6h6.R
\docType{data}
\name{c6h6}
\alias{c6h6}
\title{Tibble con i dati di c6h6 per 6 stazioni della regione TOSCANA}
\format{
Un tibble con 8 colonne e 4386 osservazioni
}
\usage{
c6h6
}
\description{
Tibble con i dati di c6h6 per 6 stazioni della regione TOSCANA
}
\keyword{datasets}
